Revert "Server endpoints (#16)"
This reverts commit b2383cd92c7ff2817764d8fe20c0d8b8d2b4b0ee.
