+++ Pathfinding/CarrierPidgeon/BusinessLayer/Types/Enums/nodeProperites.enum.cs
@@ -1,0 +1,7 @@
+namespace CarrierPidgeon.Business_Layer
+{
+    public enum NodeProperties
+    {
+        Obstacle,
+    }
+}
