+++ Pathfinding/CarrierPidgeon/BusinessLayer/Types/Enums/uniqueNodeProperties.enum.cs
@@ -1,0 +1,8 @@
+namespace CarrierPidgeon.Business_Layer
+{
+    public enum UniqueNodeProperties
+    {
+        Start,
+        End,
+    }
+}
