+++ Pathfinding/CarrierPidgeon/DataAccessLayer/Routes/dashboard.routes.cs
@@ -1,0 +1,16 @@
+namespace CarrierPidgeon
+{
+    public class Dashboard_Routes
+    {
+        public static void getDroneEndpoints(WebApplication webApplication)
+        {
+            webApplication.MapGet("/dashboard/", async () =>
+            {
+                // Return Node.origin for specific drone
+
+                return "Dashboard ";
+            });
+
+        }
+    }
+}
