+++ Pathfinding/CarrierPidgeon/DataAccessLayer/Routes/index.routes.cs
@@ -1,0 +1,11 @@
+namespace CarrierPidgeon
+{
+    public class Api_Endpoints
+    {
+        public static void getEnpoints(WebApplication webApplication)
+        {
+            Drone_Routes.getDroneEndpoints(webApplication);
+            Dashboard_Routes.getDroneEndpoints(webApplication);
+        }
+    }
+}
