+++ Pathfinding/CarrierPidgeon/PresentationLayer/Program.cs
@@ -1,0 +1,16 @@
+﻿namespace CarrierPidgeon
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            WebApplicationBuilder webApplicationBuilder = WebApplication.CreateBuilder(args);
+
+            WebApplication webApplication = webApplicationBuilder.Build();
+
+            Api_Endpoints.getEnpoints(webApplication);
+
+            webApplication.Run();
+        }
+    }
+}
