--- Pathfinding/CarrierPidgeon/Program.cs
@@ -1,19 +1,0 @@
-﻿using CarrierPidgeon.Config;
-
-namespace CarrierPidgeon
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
-}
