+++ Pathfinding/CarrierPidgeon/Properties/launchSettings.json
@@ -2,11 +2,11 @@
   "profiles": {
     "CarrierPidgeon": {
       "commandName": "Project",
-      "launchBrowser": false,
-      "applicationUrl": "http://localhost:4000",
+      "launchBrowser": true,
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
-      }
+      },
+      "applicationUrl": "https://localhost:4000;http://localhost:4000"
     }
   }
 }
