--- Pathfinding/CarrierPidgeon/src/Config/route.config.cs
@@ -1,23 +1,0 @@
-namespace CarrierPidgeon.Config;
-
-public class RouteConfig
-{
-    public static void Configure(IApplicationBuilder app, IWebHostEnvironment environment)
-    {
-        ConfigureEnvironment(app, environment);
-    }
-
-    public static void ConfigureEnvironment(IApplicationBuilder app, IWebHostEnvironment environment)
-    {
-        app.UseRouting();
-
-        if (environment.IsDevelopment())
-        {
-            app.UseDeveloperExceptionPage();
-        }
-        else
-        {
-            // Configure production settings
-        }
-    }
-}
