--- Pathfinding/CarrierPidgeon/src/Middleware/grid.middleware.cs
@@ -1,14 +1,0 @@
-// TODO
-using CarrierPidgeon.Models;
-using CarrierPidgeon.Services;
-
-namespace CarrierPidgeon.Middleware
-{
-    public class GridMiddleware
-    {
-        public static List<Node> initGrid()
-        {
-            return GridServices.InitializeNodes();
-        }
-    }
-}
