--- Pathfinding/CarrierPidgeon/src/Models/grid.model.cs
@@ -1,10 +1,0 @@
-using CarrierPidgeon.Types;
-
-namespace CarrierPidgeon.Models
-{
-    public class Grid : IGrid
-    {
-        public int sizeX {get; set;} 
-        public int sizeY {get; set;}  
-    }
-}
