--- Pathfinding/CarrierPidgeon/src/Services/drone.services.cs
@@ -1,9 +1,0 @@
-using CarrierPidgeon.Models;
-
-namespace CarrierPidgeon.Services
-{
-    public class DroneServices
-    {
-        public Drone drone {get;} = new Drone();
-    }
-}
