--- Pathfinding/CarrierPidgeon/src/Services/grid.services.cs
@@ -1,17 +1,0 @@
-using CarrierPidgeon.Models;
-
-namespace CarrierPidgeon.Services
-{
-    public class GridServices
-    {
-        public Grid grid {get;} = new Grid();
-
-        public static List<Node> InitializeNodes()
-        {
-            // Initialize the nodes array with default values
-            int width = 10;
-            int height = 10;
-            return NodeServices.initNodes(width, height);
-        }
-    }
-}
