--- Pathfinding/CarrierPidgeon/src/Types/Enums/nodeProperites.enum.cs
@@ -1,7 +1,0 @@
-namespace CarrierPidgeon.Types
-{
-    public enum NodeProperties
-    {
-        Obstacle,
-    }
-}
