--- Pathfinding/CarrierPidgeon/src/Types/Enums/uniqueNodeProperties.enum.cs
@@ -1,8 +1,0 @@
-namespace CarrierPidgeon.Types
-{
-    public enum UniqueNodeProperties
-    {
-        Start,
-        End,
-    }
-}
