--- Pathfinding/CarrierPidgeon/src/Types/Interfaces/drone.interface.cs
@@ -1,12 +1,0 @@
-using CarrierPidgeon.Models;
-using MongoDB.Bson;
-
-namespace CarrierPidgeon.Types
-{
-    public interface IDrone
-    {
-        ObjectId droneId {get; set;}
-        string status {get; set;}
-        Models.Route activeRoute {get; set;}
-    }
-}
