--- Pathfinding/CarrierPidgeon/src/Types/Interfaces/enumInfo.interface.cs
@@ -1,6 +1,0 @@
-namespace CarrierPidgeon.Types{
-    public interface IEnumInfo{
-        string type {get; set;}
-        string value {get; set;}
-    }
-}
