--- Pathfinding/CarrierPidgeon/src/Types/Interfaces/grid.interface.cs
@@ -1,8 +1,0 @@
-namespace CarrierPidgeon.Types
-{
-    public interface IGrid
-    {
-        int sizeX {get; set;}
-        int sizeY {get; set;}
-    }
-}
